Solve some issue
