+++ src/composeApp/src/commonMain/kotlin/RootApp.kt
@@ -1,6 +1,4 @@
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.LaunchedEffect
-import androidx.compose.runtime.remember
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.rememberNavController
 import com.gabrielbmoro.moviedb.details.ui.screens.details.DetailsScreen
@@ -12,14 +10,10 @@ import com.gabrielbmoro.moviedb.platform.navigation.addSearchScreen
 import com.gabrielbmoro.moviedb.platform.navigation.addWishlistScreen
 import com.gabrielbmoro.moviedb.search.ui.screens.search.SearchScreen
 import com.gabrielbmoro.moviedb.wishlist.ui.screens.wishlist.WishlistScreen
-import kotlinx.coroutines.flow.MutableStateFlow
-import kotlinx.coroutines.flow.collectLatest
 
 @Composable
 fun RootApp() {
     val navigator = rememberNavController()
-
-    val deepLinkNavigation = remember { DeepLinkNavigation() }
 
     NavHost(
         startDestination = Screen.Movies.route,
@@ -27,15 +21,6 @@ fun RootApp() {
     ) {
         addMoviesScreen {
             MoviesScreen(navigator = navigator)
-
-            // Defined deeplink barrier to the first screen destination
-            LaunchedEffect(Unit) {
-                deepLinkNavigation.barrier.collectLatest { route ->
-                    route?.let {
-                        navigator.navigate(it)
-                    }
-                }
-            }
         }
 
         addMovieDetailsScreen { movieId ->
@@ -53,10 +38,6 @@ fun RootApp() {
     }
 
     DeeplinkEffect(
-        deepLinkBarrier = deepLinkNavigation.barrier,
+        navigator = navigator,
     )
-}
-
-class DeepLinkNavigation {
-    val barrier = MutableStateFlow<String?>(null)
 }
